+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#762](https://github.com/alleslabs/celatone-frontend/pull/762) api v1 - proposal validator votes info
 - [#745](https://github.com/alleslabs/celatone-frontend/pull/745) Add proposal top
 - [#758](https://github.com/alleslabs/celatone-frontend/pull/758) api v1 - proposal validator votes
 - [#757](https://github.com/alleslabs/celatone-frontend/pull/757) api v1 - proposal data